brcmfmac: Fix transmit stalls caused by bad patch
Commit [1] changes the transmit credits policy to allow two slots for
control traffic. In doing so the author removed some careful (u8) casts
added to cope with wrap-around in an 8-bit sequence number. Removing
the casts causes the appearance of a lack of credit as soon as the
sequence number wraps back to zero.

It's staggering that this was accepted upstream.

Fixes: [1]


[1] b41c232d3366 ("brcmfmac: reserve 2 credits for host tx control path")

+++ drivers/net/wireless/broadcom/brcm80211/brcmfmac/sdio.c
@@ -668,14 +668,14 @@ static void pkt_align(struct sk_buff *p, int len, int align)
 static bool data_ok(struct brcmf_sdio *bus)
 {
 	/* Reserve TXCTL_CREDITS credits for txctl */
-	return (bus->tx_max - bus->tx_seq) > TXCTL_CREDITS &&
+	return (u8)(bus->tx_max - bus->tx_seq) > TXCTL_CREDITS &&
 	       ((bus->tx_max - bus->tx_seq) & 0x80) == 0;
 }
 
 /* To check if there's window offered */
 static bool txctl_ok(struct brcmf_sdio *bus)
 {
-	return (bus->tx_max - bus->tx_seq) != 0 &&
+	return (u8)(bus->tx_max - bus->tx_seq) != 0 &&
 	       ((bus->tx_max - bus->tx_seq) & 0x80) == 0;
 }
